OnKeyboardActionListener extracted to a different file
